pyproject.toml
[tool.poetry]
name = "weekend"
version = "0.1.0"
description = ""
authors = ["Drávai Balázs <b.dravai@cytocast.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
dataclasses-json = "^0.5.6"
python-slugify = "^6.1.1"
redis = "^4.1.4"
fastapi = "^0.75.0"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.31"
pgcli = "^3.4.0"
uvicorn = {extras = ["standard"], version = "^0.17.5"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
